Move wl_buffer creation to wlr_screencast_common.c
