+++ src/screencast/wlr_screencast_common.c
@@ -1,0 +1,58 @@
+#include "wlr_screencast_common.h"
+
+#include "wlr-screencopy-unstable-v1-client-protocol.h"
+#include "xdg-output-unstable-v1-client-protocol.h"
+#include <fcntl.h>
+#include <limits.h>
+#include <stdbool.h>
+#include <stdlib.h>
+#include <string.h>
+#include <sys/mman.h>
+#include <sys/param.h>
+#include <sys/stat.h>
+#include <sys/wait.h>
+#include <unistd.h>
+#include <wayland-client-protocol.h>
+
+#include "screencast.h"
+#include "pipewire_screencast.h"
+#include "xdpw.h"
+#include "logger.h"
+
+struct wl_buffer *wlr_create_shm_buffer(struct xdpw_screencast_instance *cast,
+		enum wl_shm_format fmt, int width, int height, int stride,
+		void **data_out) {
+	struct xdpw_screencast_context *ctx = cast->ctx;
+	int size = stride * height;
+
+	int fd = anonymous_shm_open();
+	if (fd < 0) {
+		logprint(ERROR, "wlroots: shm_open failed");
+		return NULL;
+	}
+
+	int ret;
+	while ((ret = ftruncate(fd, size)) == EINTR);
+
+	if (ret < 0) {
+		close(fd);
+		logprint(ERROR, "wlroots: ftruncate failed");
+		return NULL;
+	}
+
+	void *data = mmap(NULL, size, PROT_READ | PROT_WRITE, MAP_SHARED, fd, 0);
+	if (data == MAP_FAILED) {
+		logprint(ERROR, "wlroots: mmap failed: %m");
+		close(fd);
+		return NULL;
+	}
+
+	struct wl_shm_pool *pool = wl_shm_create_pool(ctx->shm, fd, size);
+	close(fd);
+	struct wl_buffer *buffer =
+		wl_shm_pool_create_buffer(pool, 0, width, height, stride, fmt);
+	wl_shm_pool_destroy(pool);
+
+	*data_out = data;
+	return buffer;
+}
