+++ src/screencast/wlr_screencopy.c
@@ -1,5 +1,6 @@
 #include "wlr_screencopy.h"
 #include "wlr_screencast.h"
+#include "wlr_screencast_common.h"
 
 #include "wlr-screencopy-unstable-v1-client-protocol.h"
 #include "xdg-output-unstable-v1-client-protocol.h"
@@ -43,44 +44,6 @@ void xdpw_wlr_screencopy_frame_free(struct xdpw_screencast_instance *cast) {
 		logprint(TRACE, "xdpw: xdpw_frames.screencopy_frame buffer destroyed");
 	}
 	logprint(TRACE, "wlroots: frame destroyed");
-}
-
-static struct wl_buffer *create_shm_buffer(struct xdpw_screencast_instance *cast,
-		enum wl_shm_format fmt, int width, int height, int stride,
-		void **data_out) {
-	struct xdpw_screencast_context *ctx = cast->ctx;
-	int size = stride * height;
-
-	int fd = anonymous_shm_open();
-	if (fd < 0) {
-		logprint(ERROR, "wlroots: shm_open failed");
-		return NULL;
-	}
-
-	int ret;
-	while ((ret = ftruncate(fd, size)) == EINTR);
-
-	if (ret < 0) {
-		close(fd);
-		logprint(ERROR, "wlroots: ftruncate failed");
-		return NULL;
-	}
-
-	void *data = mmap(NULL, size, PROT_READ | PROT_WRITE, MAP_SHARED, fd, 0);
-	if (data == MAP_FAILED) {
-		logprint(ERROR, "wlroots: mmap failed: %m");
-		close(fd);
-		return NULL;
-	}
-
-	struct wl_shm_pool *pool = wl_shm_create_pool(ctx->shm, fd, size);
-	close(fd);
-	struct wl_buffer *buffer =
-		wl_shm_pool_create_buffer(pool, 0, width, height, stride, fmt);
-	wl_shm_pool_destroy(pool);
-
-	*data_out = data;
-	return buffer;
 }
 
 static void wlr_frame_buffer_chparam(struct xdpw_screencast_instance *cast,
@@ -128,7 +91,7 @@ static void wlr_frame_buffer(void *data, struct zwlr_screencopy_frame_v1 *frame,
 
 	if (cast->xdpw_frames.screencopy_frame.buffer == NULL) {
 		logprint(DEBUG, "wlroots: create shm buffer");
-		cast->xdpw_frames.screencopy_frame.buffer = create_shm_buffer(cast, format, width, height,
+		cast->xdpw_frames.screencopy_frame.buffer = wlr_create_shm_buffer(cast, format, width, height,
 			stride, &cast->xdpw_frames.screencopy_frame.data);
 	} else {
 		logprint(TRACE,"wlroots: shm buffer exists");
